requirements.txt
Django==1.7.2
MySQL-python==1.2.5
argparse==1.2.1
beautifulsoup4==4.3.2
cssselect==0.9.1
djangorestframework==3.1.1
selenium==2.44.0
